Conflicts:
	lib/db/upgrade.php
	mod/scorm/lib.php
	version.php

+++ mod/choice/view.php
@@ -28,7 +28,7 @@
         print_error('badcontext');
     }
 
-    if ($action == 'delchoice') {
+    if ($action == 'delchoice' and confirm_sesskey() and has_capability('mod/choice:choose', $context) and $choice->allowupdate) {
         if ($answer = get_record('choice_answers', 'choiceid', $choice->id, 'userid', $USER->id)) {
             //print_object($answer);
             delete_records('choice_answers', 'id', $answer->id);
@@ -39,7 +39,7 @@
                   update_module_button($cm->id, $course->id, $strchoice), navmenu($course, $cm));
 
 /// Submit any new data if there is any
-    if ($form = data_submitted() && has_capability('mod/choice:choose', $context)) {
+    if ($form = data_submitted() && has_capability('mod/choice:choose', $context) && confirm_sesskey()) {
         $timenow = time();
         if (has_capability('mod/choice:deleteresponses', $context)) {
             if ($action == 'delete') { //some responses need to be deleted     
@@ -66,7 +66,7 @@
     
     if ($groupmode) {
         groups_get_activity_group($cm, true);
-        groups_print_activity_menu($cm, 'view.php?id='.$id);
+        groups_print_activity_menu($cm, $CFG->wwwroot . '/mod/choice/view.php?id='.$id);
     }
     $allresponses = choice_get_response_data($choice, $cm, $groupmode);   // Big function, approx 6 SQL calls per user
